third_party/WebKit/LayoutTests/inspector/tracing/timeline-paint-with-layout-invalidations-expected.txt

Tests the Timeline API instrumentation of paint events with layout invalidations.

PASS


Running: testLocalFrame

Running: testSubframe
